Move tests (final_final2)
