--- constraint-solver/src/tests/mod.rs
@@ -1,2 +1,0 @@
-mod solver;
-mod test_utils;
