+++ constraint-solver/tests/solver.rs
@@ -1,8 +1,10 @@
 use std::collections::BTreeMap;
 
+use powdr_constraint_solver::{
+    quadratic_symbolic_expression::QuadraticSymbolicExpression, solver::Solver,
+};
 use powdr_number::GoldilocksField;
-
-use crate::{quadratic_symbolic_expression::QuadraticSymbolicExpression, solver::Solver};
+use test_log::test;
 
 pub type Var = &'static str;
 pub type Qse = QuadraticSymbolicExpression<GoldilocksField, Var>;
@@ -48,4 +50,51 @@ fn assert_expected_state(
         }
     }
     assert!(!error, "Final state does not match expected state");
+}
+
+#[test]
+fn single_variable() {
+    assert_solve_result(vec![var("x") - constant(5)], vec![("x", 5.into())]);
+}
+
+#[test]
+fn concretely_solvable() {
+    assert_solve_result(
+        vec![
+            var("a") - constant(2),
+            var("b") - constant(3),
+            // c = a * b = 6
+            var("c") - var("a") * var("b"),
+            // d = c * 4 - a = 22
+            var("d") - (var("c") * constant(4) - var("a")),
+        ],
+        vec![
+            ("a", 2.into()),
+            ("b", 3.into()),
+            ("c", 6.into()),
+            ("d", 22.into()),
+        ],
+    );
+}
+
+#[test]
+fn bit_decomposition() {
+    assert_solve_result(
+        vec![
+            // 4 bit-constrained variables:
+            var("b0") * (var("b0") - constant(1)),
+            var("b1") * (var("b1") - constant(1)),
+            var("b2") * (var("b2") - constant(1)),
+            var("b3") * (var("b3") - constant(1)),
+            // Bit-decomposition of a concrete value:
+            var("b0") + var("b1") * constant(2) + var("b2") * constant(4) + var("b3") * constant(8)
+                - constant(0b1110),
+        ],
+        vec![
+            ("b0", 0.into()),
+            ("b1", 1.into()),
+            ("b2", 1.into()),
+            ("b3", 1.into()),
+        ],
+    );
 }
